+++ elide-spring/elide-spring-boot-autoconfigure/src/main/java/com/yahoo/elide/spring/config/ElideAutoConfiguration.java
@@ -219,7 +219,7 @@ public class ElideAutoConfiguration {
                             aggregationDataStore);
         }
 
-        return new MultiplexManager(jpaDataStore);
+        return jpaDataStore;
     }
 
     /**
