--- elide-spring/elide-spring-boot-autoconfigure/src/test/java/example/tests/DisableAggStoreIntegrationTest.java
@@ -1,36 +1,0 @@
-/*
- * Copyright 2019, Yahoo Inc.
- * Licensed under the Apache License, Version 2.0
- * See LICENSE file in project root for terms.
- */
-package example.tests;
-
-import org.junit.jupiter.api.BeforeAll;
-import org.junit.jupiter.api.TestInstance;
-import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.boot.web.server.LocalServerPort;
-import org.springframework.test.context.ActiveProfiles;
-
-import io.restassured.RestAssured;
-
-import java.util.TimeZone;
-
-/**
- * Base class for running a set of functional Elide tests.  This class
- * sets up an Elide instance with an in-memory H2 database.
- */
-@SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
-@TestInstance(TestInstance.Lifecycle.PER_CLASS)
-@ActiveProfiles("disableAggStore")
-public class DisableAggStoreIntegrationTest {
-
-    @LocalServerPort
-    protected int port;
-
-    @BeforeAll
-    public void setUp() {
-        TimeZone.setDefault(TimeZone.getTimeZone("UTC"));
-        RestAssured.port = port;
-        RestAssured.enableLoggingOfRequestAndResponseIfValidationFails();
-    }
-}
